Don't set up enclave if we're not in enclave.
To facilitate working with the code outside of a Nitro enclave, this
patch adds the function InEnclave, which determines if the code is
running in an enclave (by checking for the existence of the /dev/nsm
device).  If we're not in an enclave, we skip enclave-specific setup
tasks like assigning an IP address to the loopback device and using a
VSOCK instead of a TCP listener.