Make file descriptor limit configurable.
It's easy to hit the process's file descriptor limit if an enclave deals
with a substantial number of open network connections.  This patch
therefore makes it possible to increase the soft and hard limit by
setting it in the enclave config.  If the limits are not set, we are
using a default of 65,535 for both.