Only use certificate cache outside an enclave.
We don't need a certificate cache when we're running inside an enclave
because when the enclave is gone, the certificate cache will be gone as
well.  Besides, the certificate cache makes it more difficult for the
enclave to switch to an unprivileged user before execution because we
now have to deal with the permissions of the cache directory and its
content.

This fixes https://github.com/brave/nitriding/issues/35.