Add extra byte to account for "\n".
When an enclave initiates key synchronization, it first fetches a nonce.
Critically, the enclave does not read a single byte more than it has to:

    maxReadLen := base64.StdEncoding.EncodedLen(nonceLen)

The nonce handler however adds a "\n" to the Base64-encoded nonce, which
exceeds the read limit by one.  This commit fixes the issue by adding an
extra byte to the maximum allowable number of bytes.

(The commit also refactors the syntax of a map literal in our tests
because the latest "go fmt" seems to have changed its mind on that.)