Add command line flag to run enclave application.
This commit adds a new command line flag that instructs nitriding to run
the given enclave application.  Nitriding then blocks until the command
finishes (if ever) and prints both the command's stdout and stderr.

This feature has the advantage that one does not need a shell script to
start both nitriding and the enclave application, meaning that it's now
easier to build very slim Docker images that only contain a
statically-compiled nitriding and enclave application.

This fixes https://github.com/brave/nitriding/issues/58.