Make busy rx/tx loop faster.
This commit improves the performance of the rx and tx loop by 1) moving
allocations to outside the loop and by 2) eliminating an unnecessary
Write call.

As long as gvproxy implements the same improvements, this should result
in ~20% more reqs/sec.

For posterity, here's the upstream PR for the gvproxy side of things:
https://github.com/containers/gvisor-tap-vsock/pull/215