Add command line flags to export metrics.
This commit adds the -prometheus-port and -profile command line flags.

If set, the flag -prometheus-port instructs nitriding to expose an HTTP
server on the given port which serves Prometheus metrics.  Prometheus
metrics are coarse-grained and should therefore be safe to use in
production.  We use a separate Web server for Prometheus metrics because
that allows us to keep the endpoint private.

If set, the flag -profile instructs nitriding to expose pprof profiling
information at /enclave/debug.  Unlike Prometheus metrics, profiling
information is privacy sensitive and therefore must not be enabled in
production.  As a safeguard, this commit disables the attestation
handler if profiling is enabled.