Collapse code into a single main package.
We have been maintaining a main package, a nitriding package, and a
randseed package.  All these packages are a relic of the past.  This
commit moves all code to the root directory, so we only have to maintain
a single main package.

This has the added benefit of no longer needing the "cmd" suffix when
installing nitriding, i.e., users can now run:

    go install github.com/brave/nitriding-daemon@latest

This PR fixes https://github.com/brave/nitriding-daemon/issues/12.